+++ src/chains/WorldChain.sol
@@ -1,0 +1,135 @@
+// SPDX-License-Identifier: MIT
+pragma solidity =0.8.25;
+
+import {SettlerBase} from "../SettlerBase.sol";
+import {Settler} from "../Settler.sol";
+import {SettlerMetaTxn} from "../SettlerMetaTxn.sol";
+
+import {IERC20} from "@forge-std/interfaces/IERC20.sol";
+import {FreeMemory} from "../utils/FreeMemory.sol";
+
+import {ISettlerActions} from "../ISettlerActions.sol";
+import {ISignatureTransfer} from "@permit2/interfaces/ISignatureTransfer.sol";
+import {UnknownForkId} from "../core/SettlerErrors.sol";
+
+import {
+    uniswapV3WorldChainFactory,
+    uniswapV3InitHash,
+    uniswapV3ForkId,
+    IUniswapV3Callback
+} from "../core/univ3forks/UniswapV3.sol";
+import {dackieSwapV3WorldChainFactory, dackieSwapV3ForkId} from "../core/univ3forks/DackieSwapV3.sol";
+import {pancakeSwapV3InitHash, IPancakeSwapV3Callback} from "../core/univ3forks/PancakeSwapV3.sol";
+
+// Solidity inheritance is stupid
+import {SettlerAbstract} from "../SettlerAbstract.sol";
+import {AbstractContext} from "../Context.sol";
+import {Permit2PaymentAbstract} from "../core/Permit2PaymentAbstract.sol";
+
+abstract contract WorldChainMixin is FreeMemory, SettlerBase {
+    constructor() {
+        assert(block.chainid == 480 || block.chainid == 31337);
+    }
+
+    function _dispatch(uint256 i, uint256 action, bytes calldata data)
+        internal
+        virtual
+        override(SettlerBase)
+        DANGEROUS_freeMemory
+        returns (bool)
+    {
+        if (super._dispatch(i, action, data)) {
+            return true;
+        } else {
+            return false;
+        }
+        return true;
+    }
+
+    function _uniV3ForkInfo(uint8 forkId)
+        internal
+        pure
+        override
+        returns (address factory, bytes32 initHash, uint32 callbackSelector)
+    {
+        if (forkId == uniswapV3ForkId) {
+            factory = uniswapV3WorldChainFactory;
+            initHash = uniswapV3InitHash;
+            callbackSelector = uint32(IUniswapV3Callback.uniswapV3SwapCallback.selector);
+        } else if (forkId == dackieSwapV3ForkId) {
+            factory = dackieSwapV3WorldChainFactory;
+            initHash = pancakeSwapV3InitHash;
+            callbackSelector = uint32(IPancakeSwapV3Callback.pancakeV3SwapCallback.selector);
+        } else {
+            revert UnknownForkId(forkId);
+        }
+    }
+}
+
+/// @custom:security-contact security@0x.org
+contract WorldChainSettler is Settler, WorldChainMixin {
+    constructor(bytes20 gitCommit) Settler(gitCommit) {}
+
+    function _dispatchVIP(uint256 action, bytes calldata data) internal override DANGEROUS_freeMemory returns (bool) {
+        if (super._dispatchVIP(action, data)) {
+            return true;
+        } else {
+            return false;
+        }
+        return true;
+    }
+
+    // Solidity inheritance is stupid
+    function _isRestrictedTarget(address target)
+        internal
+        pure
+        override(Settler, Permit2PaymentAbstract)
+        returns (bool)
+    {
+        return super._isRestrictedTarget(target);
+    }
+
+    function _dispatch(uint256 i, uint256 action, bytes calldata data)
+        internal
+        override(SettlerAbstract, SettlerBase, WorldChainMixin)
+        returns (bool)
+    {
+        return super._dispatch(i, action, data);
+    }
+
+    function _msgSender() internal view override(Settler, AbstractContext) returns (address) {
+        return super._msgSender();
+    }
+}
+
+/// @custom:security-contact security@0x.org
+contract WorldChainSettlerMetaTxn is SettlerMetaTxn, WorldChainMixin {
+    constructor(bytes20 gitCommit) SettlerMetaTxn(gitCommit) {}
+
+    function _dispatchVIP(uint256 action, bytes calldata data, bytes calldata sig)
+        internal
+        override
+        DANGEROUS_freeMemory
+        returns (bool)
+    {
+        if (super._dispatchVIP(action, data, sig)) {
+            return true;
+        } else {
+            return false;
+        }
+        return true;
+    }
+
+    // Solidity inheritance is stupid
+    function _dispatch(uint256 i, uint256 action, bytes calldata data)
+        internal
+        override(SettlerAbstract, SettlerBase, WorldChainMixin)
+        returns (bool)
+    {
+        return super._dispatch(i, action, data);
+    }
+
+    function _msgSender() internal view override(SettlerMetaTxn, AbstractContext) returns (address) {
+        return super._msgSender();
+    }
+}